+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 0.7.1 (unreleased)
 
-- Added `create` method for `Solver`
 - Added support for time limit for `Solver`
 - Added `enable_output` and `suppress_output` to `Solver`
+- Improved `new` method for `Solver`
 - Fixed error with offset with `Solver`
 - Fixed segfault with `CpSolver`